moves bridgeBase and IBridge/IBridgeAdvanced
