+++ essentials-framework/Essentials Core/PepperDashEssentialsBase/Bridges/IBridge.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using Crestron.SimplSharp;
 using Crestron.SimplSharpPro.DeviceSupport;
 
-namespace PepperDash.Essentials.Bridges
+namespace PepperDash.Essentials.Core.Bridges
 {
     /// <summary>
     /// Defines a device that uses the legacy JoinMapBase for its join map
